+++ Cartfile
@@ -1,2 +1,1 @@
-github "antitypical/Result" ~> 4.0
-github "Thomvis/BrightFutures" ~> 7.0.0
+github "Thomvis/BrightFutures" ~> 8.0.0
