+++ Tests/Helper/Mocks/MockNetworkService.swift
@@ -8,7 +8,6 @@
 
 import Foundation
 import Hyperspace
-import Result
 
 class MockNetworkService {
     private(set) var executeCallCount = 0
